+++ frontend/components/FormSentMessage.js
@@ -1,0 +1,15 @@
+
+export default function FormSentMessage(props) {
+    return (
+        <div className='section'>
+            <div className='submitPage' style={{ backgroundColor: "orange" }}>
+                <div className="stepTitle">5. Your budget</div>
+                <div className='step'>
+                    <h2>Form sent</h2>
+                    <h1>£2000</h1>
+                </div>
+            </div>
+        </div>
+    )
+
+}
